feat: não permitir deletar moto alugada

+++ src/API/MotoHub.Application/UseCases/Motorcycles/DeleteMotorcycleUseCase.cs
@@ -5,7 +5,7 @@ using MotoHub.Domain.Entities;
 
 namespace MotoHub.Application.UseCases.Motorcycles;
 
-public class DeleteMotorcycleUseCase(IMotorcycleRepository motorcycleRepository) : IDeleteMotorcycleUseCase
+public class DeleteMotorcycleUseCase(IMotorcycleRepository motorcycleRepository, IRentRepository rentRepository) : IDeleteMotorcycleUseCase
 {
     public async Task<Result> ExecuteAsync(string identifier, CancellationToken cancellationToken = default)
     {
@@ -14,6 +14,13 @@ public class DeleteMotorcycleUseCase(IMotorcycleRepository motorcycleRepository)
         if (motorcycle is null)
         {
             return Result.Failure("Moto não encontrada", ResultErrorType.NotFound);
+        }
+
+        bool isMotorcycleRented = await rentRepository.IsMotorcycleCurrentlyRentedAsync(motorcycle.Id, cancellationToken);
+
+        if (isMotorcycleRented)
+        {
+            return Result.Failure("Moto está atualmente alugada", ResultErrorType.BusinessError);
         }
 
         await motorcycleRepository.DeleteAsync(motorcycle.Id, cancellationToken);
